requirements.txt
azure-functions
matplotlib
matplotlib-venn
numpy
grpcio
